fix warning

+++ stepmania/src/OptionRowHandler.cpp
@@ -26,7 +26,7 @@
 
 static void SelectExactlyOne( int iSelection, vector<bool> &vbSelectedOut )
 {
-	ASSERT_M( iSelection >= 0  &&  iSelection < vbSelectedOut.size(), ssprintf("%d/%u",iSelection,vbSelectedOut.size()) );
+	ASSERT_M( iSelection >= 0  &&  iSelection < (int) vbSelectedOut.size(), ssprintf("%d/%u",iSelection,vbSelectedOut.size()) );
 	for( int i=0; i<(int)vbSelectedOut.size(); i++ )
 		vbSelectedOut[i] = i==iSelection;
 }
